inclusão de botões para desativar a vibração e o som

+++ app/src/main/java/br/com/cristiano/etanolougasolina/aux/VerificadorCombustivel.java
@@ -13,7 +13,7 @@ public class VerificadorCombustivel implements VerificarCombustivel {
     }
 
     @Override
-    public Boolean ehGasolina(double gasolina, double etanol) {
+    public Boolean ehGasolina(Float gasolina, Float etanol) {
         return (etanol / gasolina) > TAXA;
     }
 }
