+++ app/src/main/java/br/com/cristiano/etanolougasolina/interfaces/VerificarCombustivel.java
@@ -2,5 +2,5 @@ package br.com.cristiano.etanolougasolina.interfaces;
 
 public interface VerificarCombustivel {
 
-    Boolean ehGasolina(double gasolina, double etanol);
+    Boolean ehGasolina(Float gasolina, Float etanol);
 }
